+++ cmem_plugin_shapes/plugin_shapes.py
@@ -180,10 +180,13 @@ class ShapesPlugin(WorkflowPlugin):
         self.prefix_cc = prefix_cc
 
         self.ignore_properties = []
-        for _ in ignore_properties.split("\n"):
-            if not validators.url(_):
-                raise ValueError(f"Invalid property IRI ({_}) in parameter 'Properties to ignore'")
-            self.ignore_properties.append(_)
+        if ignore_properties.strip():
+            for _ in ignore_properties.split("\n"):
+                if not validators.url(_):
+                    raise ValueError(
+                        f"Invalid property IRI ({_}) in parameter 'Properties to ignore'"
+                    )
+                self.ignore_properties.append(_)
 
         self.plugin_provenance = plugin_provenance