+++ cmem_plugin_shapes/plugin_shapes.py
@@ -119,9 +119,10 @@ def str2bool(value: str) -> bool:
             param_type=BoolParameterType(),
             name="prefix_cc",
             label="Fetch namespace prefixes from prefix.cc",
-            description="Attempt to fetch namespace prefixes from prefix.cc instead of from the "
-            "local database. If this fails, fall back on local database. Prefixes defined in the "
-            "Corporate Memory project will override prefixes defined in the external database.",
+            description="Attempt to fetch the list of namespace prefixes from https://prefix.cc "
+            "instead of using the local database. If this fails, fall back on the local database. "
+            "Prefixes defined in the Corporate Memory project will override prefixes defined in "
+            "the database.",
             advanced=True,
         ),
         PluginParameter(