+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## Unreleased
+
+### Fixed
+
+- Fixed error when leaving "Properties to ignore" empty
 
 ## [3.0.0] 2025-03-05