+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Fixed
 
-- Fixed error when leaving "Properties to ignore" empty
+- Fixed error when "Properties to ignore" field is empty (or contains empty lines) 
 
 ## [3.0.0] 2025-03-05